+++ R/to_as4wrf.R
@@ -52,7 +52,7 @@
 to_as4wrf <- function(sdf, nr = 1, dmyhm, tz, crs = 4326, islist){
   if(nr <= 0){
     stop("The argument 'nr' must be positive") # nocov
-  } else if (class(sdf)[1] != "list") {
+  } else if (!inherits(sdf, "list")) {
     sdf <- sf::st_as_sf(sdf)
     # if(class(sdf)[1] == "sf"){
     #   sdf <- as(sdf, "Spatial")
@@ -88,7 +88,7 @@ to_as4wrf <- function(sdf, nr = 1, dmyhm, tz, crs = 4326, islist){
         strftime(dft$time_utc, timezone = tz, format = "%d"),
         strftime(dft$time_utc, timezone = tz, format = "%H")
       ))
-  } else if (class(sdf) == "list") {
+  } else if (inherits(sdf, "list")) {
     # if(class(sdf)[1] == "sf"){
     #   sdf <- lapply(sdf, methods::as, "Spatial")
     # }