+++ R/wrf_get.R
@@ -5,16 +5,25 @@
 #' @param file name of file interactively (default) or specified
 #' @param name name of the variable (any variable) or time to return a POSIXlt object from model
 #' @param as_raster return a raster instead of an array
-#' @param raster_crs crs to use if as_raster is TRUE
+#' @param raster_crs crs of outputif as_raster is TRUE, see details
 #' @param raster_lev level for rasters from a 4D variable
+#' @param k multiplier
 #' @param verbose display additional information
+#' @param ... additional parameters passed to wrf_raster
+#'
+#' @details wrf_get can return a raster object with the option as_raster = TRUE,
+#' raster_crs can be used to specify the output crs of the raster object,
+#' raster_crs = 'latlon' can be especifyed to use latlon option in wrf_raster.
+#' If raster_crs is 'WRF' (default), the output projection is equivalent to
+#' the WRF grid.
 #'
 #' @format array or raster object
 #'
-#' @importFrom  ncdf4 nc_open nc_close ncvar_get ncatt_get
+#' @importFrom ncdf4 nc_open nc_close ncvar_get ncatt_get
 #' @importFrom raster raster brick flip
 #' @importFrom sp CRS
 #'
+#' @return Read vars from NetCDF WRF file.
 #' @export
 #'
 #' @author Daniel Schuch
@@ -51,9 +60,11 @@
 wrf_get <- function(file = file.choose(),
                     name = NA,
                     as_raster = FALSE,
-                    raster_crs = "+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs",
+                    raster_crs = 'WRF',
                     raster_lev = 1,
-                    verbose = FALSE){
+                    k = NA,
+                    verbose = FALSE, ...){
+
   if(!is.na(name)){
     if(name == 'time'){
       wrfchem <- ncdf4::nc_open(file)                                                     # nocov
@@ -66,60 +77,52 @@ wrf_get <- function(file = file.choose(),
       return(TIME)                                                                        # nocov
     }
   }
-  if(verbose)
-    cat(paste0('reading ',name,' from ', file,'\n'))                     # nocov
+  if(verbose){
+    if(missing(k)){                                                       # nocov
+      cat(paste0('reading ',name,' from ', file,'\n'))                    # nocov
+    }else{
+      cat(paste0('reading ',name,' from ', file,' k = ',k,'\n'))          # nocov
+    }
+  }
 
   wrfchem <- ncdf4::nc_open(file)                                       # iteractive
   if(is.na(name)){                                                      # nocov start
     name  <- menu(names(wrfchem$var), title = "Choose the variable:")
     POL   <- ncdf4::ncvar_get(wrfchem, names(wrfchem$var)[name])
     name  <- names(wrfchem$var)[name]                                   # nocov end
-  }else{
-    POL   <- ncvar_get(wrfchem,name)
   }
   if(as_raster){
-    if(length(dim(POL)) >= 5)                                                  # nocov start
-      stop("images with 5D or more not suported")
-
-    if(length(dim(POL)) == 4){
-      cat(paste0("4D images not supported, making a 3D RasterBrick using level ",
-                 raster_lev," of the file\n"))
-      POL <- POL[,,raster_lev,,drop = TRUE]
-    }                                                                          # nocov end
-
-    lat    <- ncdf4::ncvar_get(wrfchem, varid = "XLAT")
-    lon    <- ncdf4::ncvar_get(wrfchem, varid = "XLONG")
-    time   <- ncdf4::ncvar_get(wrfchem, varid = "Times")
-    r.lat  <- range(lat)
-    r.lon  <- range(lon)
-    n.lat  <- ncdf4::ncatt_get(wrfchem, varid = 0,
-                               attname = "SOUTH-NORTH_PATCH_END_UNSTAG")$value
-    n.lon  <- ncdf4::ncatt_get(wrfchem, varid = 0,
-                               attname = "WEST-EAST_PATCH_END_UNSTAG")$value
-
-    n      <- length(time)
-    if(n == 1){
-      r <- raster::raster(x = t(POL),
-                          xmn=r.lon[1],
-                          xmx=r.lon[2],
-                          ymn=r.lat[1],
-                          ymx=r.lat[2])
-      r <- raster::flip(r,2)
+    if(raster_crs == 'latlon' |
+       raster_crs == 'lonlat' |
+       raster_crs == 'latlong'|
+       raster_crs == 'longlat'){
+      r <- wrf_raster(file    = file,           # nocov
+                      name    = name,           # nocov
+                      latlon  = TRUE,           # nocov
+                      level   = raster_lev,     # nocov
+                      verbose = verbose,        # nocov
+                      ...)                      # nocov
+    }else{
+      r <- wrf_raster(file    = file,
+                      name    = name,
+                      latlon  = FALSE,
+                      level   = raster_lev,
+                      verbose = verbose,
+                      ...)
+      if(raster_crs != 'WRF'){
+        r <- projectRaster(r, crs=raster_crs)
+      }
     }
-    if(n > 1){                                        # for emissions in 2D+time
-      r <- raster::brick(x = aperm(POL, c(2, 1, 3)),  # nocov start
-                         xmn = r.lon[1],
-                         xmx = r.lon[2],
-                         ymn = r.lat[1],
-                         ymx = r.lat[2])
-      r <- raster::flip(r,2)                          # nocov end
+    if(!missing(k)){
+      r <- k * r                                     # nocov
     }
-    raster::crs(r)   <- sp::CRS(raster_crs)
-    names(r) <- paste(name,time,sep="_")
-    ncdf4::nc_close(wrfchem)
     return(r)
   } else {
+    POL   <- ncvar_get(wrfchem,name)
     ncdf4::nc_close(wrfchem)
+    if(!missing(k)){
+      POL <- k * POL                                 # nocov
+    }
     return(POL)
   }
 }