+++ R/wrf_grid.R
@@ -11,6 +11,7 @@
 #' @importFrom raster raster rasterToPolygons
 #' @importFrom sf st_as_sf
 #' @note The default crs is 4326 (see http://spatialreference.org/ref/epsg/)
+#' @return Returns a sf polygon grid from a NetCDF wrf file
 #' @export
 #' @examples {
 #' # Do not run
@@ -23,7 +24,7 @@ wrf_grid <- function(filewrf,
                       type = "wrfinput",
                       matrix = FALSE,
                       as_raster = FALSE){
-  cat(paste("using grid info from:", filewrf, "\n"))
+  message(paste("using grid info from:", filewrf, "\n"))
   wrf <- ncdf4::nc_open(filewrf)
   if(type == "wrfinput"){
     lat    <- ncdf4::ncvar_get(wrf, varid = "XLAT")
@@ -39,8 +40,8 @@ wrf_grid <- function(filewrf,
                              attname = "SOUTH-NORTH_PATCH_END_UNSTAG")$value
   n.lon  <- ncdf4::ncatt_get(wrf, varid = 0,
                              attname = "WEST-EAST_PATCH_END_UNSTAG")$value
-  cat(paste0("Number of lat points ", n.lat, "\n"))
-  cat(paste0("Number of lon points ", n.lon, "\n"))
+  message(paste0("Number of lat points ", n.lat, "\n"))
+  message(paste0("Number of lon points ", n.lon, "\n"))
   ncdf4::nc_close(wrf)
   # r.lat  <- range(lon)
   # r.lon  <- range(lat)
@@ -51,7 +52,7 @@ wrf_grid <- function(filewrf,
     # EM  <- matrix(0, nrow = n.lon, ncol = n.lat)
   # }
 
-  if (matrix == T){
+  if (matrix){
     return(EM)
   }