+++ R/wrf_profile.R
@@ -13,11 +13,12 @@
 #'
 #' @author Daniel Schuch
 #'
+#' @return a vector with a profile
 #' @export
 #' @note It might be deprecatedin future release
 #' @seealso \code{\link{wrf_create}} and \code{\link{to_wrf}}
 #'
-#' @examples \dontrun{
+#' @examples \donttest{
 #'
 #' # Profile based on Sao Paulo tunnel experiments
 #' data(rawprofile)
@@ -54,7 +55,10 @@
 #'      at = 0.5 + c(0, 6, 12, 18, 24),
 #'     labels = c("00:00","06:00","12:00","18:00", "00:00"))
 #'}
-wrf_profile <- function(x,file,adjust = 0 ,verbose = T){
+wrf_profile <- function(x,
+                        file,
+                        adjust = 0 ,
+                        verbose = TRUE){
   x       <- as.data.frame(x)
   times   <- wrf_get(file,"Times")
   profile <- vector(mode = "numeric",length = length(times))